+++ CCSystem/src/main/java/com/team08/CCSystem/model/Medication.java
@@ -23,7 +23,7 @@ public class Medication {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(name = "name", nullable = false, unique = true)
 	private String name;
@@ -36,7 +36,7 @@ public class Medication {
 	 * @param name
 	 * @param description
 	 */
-	public Medication(Integer id, String name, String description) {
+	public Medication(Long id, String name, String description) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -50,11 +50,11 @@ public class Medication {
 		super();
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
